==> Chapter4-Experiment2/2_MLP_Cleaning.ipynb <==
#basic package
import tqdm
import os
import numpy as np
import glob
import pandas as pd
import sys
import time
import datetime as dt
import itertools
import re
import math
from collections import Counter
import operator

#scaling
from sklearn.preprocessing import StandardScaler, MinMaxScaler

#modelling
from catboost import CatBoostClassifier, Pool

#plot
import matplotlib.pyplot as plt
import seaborn as sns

PACKAGE_PARENT = '../'
SCRIPT_DIR = os.path.dirname(os.path.realpath(os.path.join(os.getcwd(), os.path.expanduser('__file__'))))
sys.path.append(os.path.normpath(os.path.join(SCRIPT_DIR, PACKAGE_PARENT)))
from UTILS import preprocessing_Origins, is_day, is_WG_open, openDevice, cleaning_processing, FB_process, FB_daily
import config_exp2 as config

# Define parameters

print('change the configuration file if not done yet!')
path_extracted_data = config.path_extracted_data
path_initial_data = config.path_initial_data
id_run = config.id_run
dico_matching = config.dico_matching
nbr_maxdur2beremoved = config.nbr_maxdur2beremoved
dico_garden_opening_hour = config.dico_garden_opening_hour
date_first_opening_WG = config.date_first_opening_WG
print(id_run)
path_extracted_data_visual = os.path.join(path_extracted_data,'visual')
pal_ = config.pal_
close_dates = config.close_dates

#create a director if not existing
if not os.path.exists(path_extracted_data_visual):
    os.makedirs(path_extracted_data_visual)

## Preprocess - add HenID,PenID

START_TIME = time.perf_counter()
p = glob.glob(os.path.join(path_initial_data, 'Barn 4 Pen*\TagUpdates\log*'))
#print('There is %d log files:\n  %s'%(len(p),'  \n  '.join(p)))
df = preprocessing_Origins(p, config, save=True, dodevice=True)
print(df.shape)
display(df.head(2))
display(df.tail(2))
END_TIME = time.perf_counter()
print ("Total running time: %.2f mn" %((END_TIME-START_TIME)/60))
#--> saves a datafarme named: _PreprocessRecords_forcleaning

# removing <60 sec

#for being more reproductible, we open the file that was saved from cleaning
df = pd.read_csv(os.path.join(path_extracted_data, id_run+'_PreprocessRecords_forcleaning.csv'), sep=';', 
                 parse_dates=['Timestamp', 'date'], dayfirst=True) #index_col=0) 
#Note: small check that hens are indeed changin pens in the system. we know from last experiemtn that a tag cant be peaked by another
#tracking system than its one. we have also no way to verify that an animal is picked by a specific pen from a specific 
#tracking system, but lets at least verify if the tracking system is the correct one after we have the daily variable 
#as thats when we can add it, and check on the used days only
display(df[df['HenID']=='hen_1']['system'].value_counts())
df = df.drop_duplicates(subset=['Timestamp','HenID'], keep='last')
print(min(df['date'].tolist()), max(df['date'].tolist()))
print(df.shape)
df.head(3)

#add duration previous/next zones
li_df = []
#more efficient to do it per hen, as it wont need to search in the whole dataframe, and we can simply shift the timestamp column
for i, df_hen in tqdm.tqdm(df.groupby(['HenID'])):
    #as the next record date (sort by date, then simply shift by one row and add nan at then end)
    df_hen = df_hen.sort_values(['Timestamp'], ascending=True)
    
    #compute duration 
    df_hen['next_record_date'] = df_hen['Timestamp'].tolist()[1:]+[pd.NaT]
    df_hen['duration'] = df_hen.apply(lambda x: (x['next_record_date']-x['Timestamp']).total_seconds(), axis=1)

    #add next & previous record zone
    df_hen['next_zone'] = df_hen['Zone'].tolist()[1:]+[np.nan]
    df_hen['previous_zone'] = [np.nan]+df_hen['Zone'].tolist()[0:-1]

    li_df.append(df_hen)
#put again in one dataframe
df = pd.concat(li_df)
df.sort_values(['Timestamp'], ascending=True, inplace=True)
df = df.drop(['log_file_name'], axis=1)
#df.drop_duplicates(subset=['Timestamp','TagID'], keep='last', inplace=True)
df['is_dur_smaller_60sec'] = df['duration'].map(lambda x: x<60)
display(df['is_dur_smaller_60sec'].value_counts(normalize=True))
print(df.shape)
df.head(3)

#small verification
df_hen[['Timestamp','HenID','Zone','previous_zone']].head(10)

df.groupby(['Zone','is_dur_smaller_60sec'])['Timestamp'].count().reset_index()

display(df.groupby(['PenID','Zone','is_dur_smaller_60sec'])['Timestamp'].count().reset_index().head(40))
df.groupby(['PenID','Zone','is_dur_smaller_60sec'])['Timestamp'].count().reset_index().tail(40)

df.groupby(['PenID','is_dur_smaller_60sec'])['Timestamp'].count().reset_index()

#remove duration less than 1mn
df = df[~df['is_dur_smaller_60sec']]

# remove consecutive equal zone (i.e. not same timestamps but same zone)

#recompute next zone nw that it maybe changed
#add duration previous/next zones
li_df = []
#more efficient to do it per hen, as it wont need to search in the whole dataframe, and we can simply shift the timestamp column
for i, df_hen in tqdm.tqdm(df.groupby(['HenID'])):
    #as the next record date (sort by date, then simply shift by one row and add nan at then end)
    df_hen = df_hen.sort_values(['Timestamp'], ascending=True)
    
    #compute duration 
    df_hen['next_record_date'] = df_hen['Timestamp'].tolist()[1:]+[pd.NaT]
    df_hen['duration'] = df_hen.apply(lambda x: (x['next_record_date']-x['Timestamp']).total_seconds(), axis=1)

    #add next & previous record zone
    df_hen['next_zone'] = df_hen['Zone'].tolist()[1:]+[np.nan]
    df_hen['previous_zone'] = [np.nan]+df_hen['Zone'].tolist()[0:-1]

    li_df.append(df_hen)
#put again in one dataframe
df = pd.concat(li_df)
df.sort_values(['Timestamp'], ascending=True, inplace=True)
#df.drop_duplicates(subset=['Timestamp','TagID'], keep='last', inplace=True)
df['is_dur_smaller_60sec'] = df['duration'].map(lambda x: x<60)
display(df['is_dur_smaller_60sec'].value_counts(normalize=True))
print(df.shape)
df.head(3)

#if the previous zone is the same, then its cez that should be removed
df['correction_is_consecutive_equal_initial_zone'] = False
df.loc[df['previous_zone']==df['Zone'], 'correction_is_consecutive_equal_initial_zone'] = True
display(df['correction_is_consecutive_equal_initial_zone'].value_counts(normalize=True))

#small verification
df[df['HenID']=='hen_82'].sort_values(['Timestamp'])[['Timestamp','HenID','Zone','previous_zone','correction_is_consecutive_equal_initial_zone']].head(10)

df[df['HenID']=='hen_99'].sort_values(['Timestamp'])[['Timestamp','HenID','Zone','previous_zone','correction_is_consecutive_equal_initial_zone']].head(10)

#remove consecutive equal zones = TRUE
df = df[~(df['correction_is_consecutive_equal_initial_zone'])]
print(df.shape)
df.head(3)

df[df['HenID']=='hen_99'].sort_values(['Timestamp'])[['Timestamp','HenID','Zone','correction_is_consecutive_equal_initial_zone']].head(10)

## Remove records in WG when WG was close

#remove last day (we did HA) (we remove now as WG info is not available on taht day)
df = df[df['date']!=dt.datetime(2022,7,5)]

df['is_WG_open'] = df['Timestamp'].map(lambda x: is_WG_open(x, config, epsi_open=0, epsi_close=20))
#we keep the epsi_open=0 as the system appeared to always open between 20sec of 2mn AFTER the indicated time (from video 
#observation)
#we set epsi_close=20 as for video observation we saw that itwould be between + and -15 mn than the indicated time. So we added
#some flexibility by puting 20mn.
print(df['is_WG_open'].value_counts(normalize=True))
display(df[(df['Zone']=='1_Zone')&(df['is_WG_open'])][['Timestamp','is_WG_open']].head(5))
display(df[(df['Zone']=='1_Zone')&(~df['is_WG_open'])][['Timestamp','is_WG_open']].head(5))

#small verification
df[(~df['is_WG_open'])&(df['Zone']=='1_Zone')].head(5)

#remove WG when close
print(df.shape)
df = df[~((~df['is_WG_open'])&(df['Zone']=='1_Zone'))]
print(df.shape)

# remove transitions smaller than 15mn over night

#compute new duration
li_df = []
for i, df_hen in tqdm.tqdm(df.groupby(['HenID'])):
    df_hen = df_hen.sort_values(['Timestamp'], ascending=True)
    df_hen['next_record_date'] = df_hen['Timestamp'].tolist()[1:]+[np.nan]
    df_hen['duration'] = df_hen.apply(lambda x: (x['next_record_date']-x['Timestamp']).total_seconds(), axis=1)
    li_df.append(df_hen)
df = pd.concat(li_df)
df['is_day'] = df['Timestamp'].map(lambda x: is_day(x, config.dico_night_hour))
print(df.shape)
display(df.head(3))

#small verification check all what we remove
df[((df['duration']<(15*60)) & (df['is_day']==False))][['HenID','Timestamp','is_day','duration','next_record_date']].tail(3)

#remove
df = df[~((df['duration']<(15*60)) & (df['is_day']==False))]
print(df.shape)
df.head(3)

# Save

print(df.shape)
df.to_csv(os.path.join(path_extracted_data, id_run+'_CLEANEDDATA.csv'), sep=';', index=False)
